the whole project compiles on linux

+++ Infrastructure/src/GLDevice.cpp
@@ -650,7 +650,7 @@ void GLDevice::vm_applySamplerTexture2D(GPUHandle* sampler, GPUHandle* texture, 
     applyFilter2D(filter,mipmaps,GL_TEXTURE_2D);
     glTexParameteri(GL_TEXTURE_2D,GL_TEXTURE_WRAP_S,getAddressMode(U));
     glTexParameteri(GL_TEXTURE_2D,GL_TEXTURE_WRAP_T,getAddressMode(V));
-    float border[4] = {border_color.R,border_color.G,border_color.B,border_color.A};
+    float border[4] = {(float)border_color.R,(float)border_color.G,(float)border_color.B,(float)border_color.A};
     glTexParameterfv(GL_TEXTURE_2D,GL_TEXTURE_BORDER_COLOR,border);
     if(func != CompareFunction::NEVER)
         glTexParameteri(GL_TEXTURE_2D,GL_TEXTURE_COMPARE_FUNC,getCompareFunc(func));
